requirements.txt
langchain
langchain-groq
pandas 
streamlit
python-dotenv
altair==4.2.2